Moved map loading to Content.
